+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Enhanced PWM channel fault detection logging with detailed voltage and current readings
 - Added optional raw data output to ADCHelper::printDebug() for debugging ADC readings
 - Added configuration settings for motor and water temperature sensors (NONE, EXTERNAL, and DS18B20)
+- Added TDS calibration offset configuration options for product and brine salinity sensors. (fixes #5)
 
 ## 🐛 Bug Fixes